code: teminal と terminal controller を分割。
